Split mk2 tests to workbench and actual interface difference tests
